add search history
