+++ src/screens/OfferingViewing/Components/History/SearchHistory.jsx
@@ -1,0 +1,11 @@
+import React, { useState } from 'react'
+import { Link } from 'react-router-dom'
+import { util, api } from 'utils'
+
+export default function SearchHistory({ searchHistory, offerings, removeSearchHistory }) {
+  return (
+    <div className="watch-history">
+      <h1 className="history-title">Search History</h1>
+    </div>
+  )
+}
