+++ src/screens/OfferingViewing/Components/History/index.css
@@ -1,14 +1,17 @@
-.watch-history {
+.history-bg {
   padding-top: 2em;
   width: 100%;
 }
-.watch-history .watch-history-title {
+.history-bg .watch-history {
+  width: 100%;
+}
+.history-bg .watch-history .history-title {
   margin-top: 10px;
   font-size: 1.2em;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
 }
 
-.watch-history .watch-history-item {
+.history-bg .watch-history-item {
   position: relative;
   padding: 1em 0;
   width: 100%;
@@ -17,21 +20,21 @@
   flex-direction: row;
 }
 
-.watch-history .watch-history-item:hover .media-info .media-name, .watch-history .watch-history-item:focus .media-info .media-name {
+.history-bg .watch-history-item:hover .media-info .media-name, .history-bg .watch-history-item:focus .media-info .media-name {
   color: rgb(53, 138, 141);
 }
 
-.watch-history .watch-history-item .media-info {
+.history-bg .watch-history-item .media-info {
   margin-left: 1.5em;
   margin-top: 1em;
 }
-.watch-history .watch-history-item .media-info .media-name {
+.history-bg .watch-history-item .media-info .media-name {
   font-family: Arial, Helvetica, sans-serif;
   font-weight: bold;
   font-size: 1.3em;
   color: rgb(0, 0, 0);
 }
-.watch-history .watch-history-item .media-info .offering-num {
+.history-bg .watch-history-item .media-info .offering-num {
   font-size: 1.1em;
   color: rgb(104, 104, 104);
 }
@@ -44,7 +47,7 @@
 
 
 /* Close Button */
-.watch-history .watch-history-item-con .button[type=remove-history] {
+.history-bg .watch-history-item-con .button[type=remove-history] {
   position: absolute;
   color: rgb(114, 114, 114);
   background: transparent;
